+++ src/main/java/lotto/domain/Prize.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public enum Prize {
     NOTHING(0L),
