+++ src/test/java/lotto/domain/BudgetTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
