+++ src/test/java/lotto/domain/NumbersTest.java
@@ -1,9 +1,10 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Set;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+
 
 public class NumbersTest {
