+++ src/test/java/lotto/domain/PrizeCalculatorTest.java
@@ -1,10 +1,11 @@
-package lotto;
+package lotto.domain;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+
 
 public class PrizeCalculatorTest {
